+++ src/mo_config/args.ml
@@ -40,8 +40,9 @@ let inclusion_args = [
     \      <file>"
   ]
 
-let ai_errors_args = [
+let ai_args = [
   "--ai-errors", Arg.Set Flags.ai_errors, " emit AI tailored errors";
+  "--implicit-package", Arg.String (fun s -> Flags.implicit_package := Some s), _UNDOCUMENTED_ " allow contextual dot and implicits resolution from all modules in the given package"
 ]
 
 let persistent_actors_args = [