+++ src/mo_config/flags.ml
@@ -34,6 +34,7 @@ let dump_tc = ref false
 let dump_lowering = ref false
 let check_ir = ref true
 let package_urls : string M.t ref = ref M.empty
+let implicit_package : string option ref = ref None
 let actor_aliases : string M.t ref = ref M.empty
 let actor_idl_path : string option ref = ref None
 let max_stable_pages_default = 65536
@@ -82,7 +83,10 @@ let experimental_stable_memory = ref experimental_stable_memory_default
 let typechecker_combine_srcs = ref false (* useful for the language server *)
 
 let default_warning_levels = M.empty
-  |> M.add "M0223" Allow
+  |> M.add "M0223" Allow (* don't report redundant instantions *)
+  |> M.add "M0235" Allow (* don't deprecate for non-caffeine *)
+  |> M.add "M0236" Allow (* don't suggest contextual dot notation *)
+  |> M.add "M0237" Allow (* don't report redundant explicit arguments *)
 
 let warning_levels = ref default_warning_levels