+++ src/mo_def/syntax.ml
@@ -211,6 +211,7 @@ and exp' =
   | AssertE of assert_kind * exp               (* assertion *)
   | AnnotE of exp * typ                        (* type annotation *)
   | ImportE of (string * resolved_import ref)  (* import statement *)
+  | ImplicitLibE of string                     (* implicitly imported library *)
   | ThrowE of exp                              (* throw exception *)
   | TryE of exp * case list * exp option       (* catch exception / finally *)
   | IgnoreE of exp                             (* ignore *)
@@ -318,7 +319,13 @@ let string_of_lit = function
   | FloatLit f    -> Numerics.Float.to_pretty_string f
   | PreLit _      -> assert false
 
-
+(** Used for debugging *)
+let string_of_resolved_import = function
+  | LibPath {path; package } -> Printf.sprintf "LibPath {path = %s, package = %s}" path (match package with | Some p -> p | None -> "None")
+  | IDLPath (path, _) -> Printf.sprintf "IDLPath (%s, _)" path
+  | ImportedValuePath path -> Printf.sprintf "ImportedValuePath %s" path
+  | PrimPath -> "PrimPath"
+  | Unresolved -> "Unresolved"
 
 (* Miscellaneous *)
 (* TODO: none of what follows should probably be in this file *)