+++ src/mo_frontend/typing.ml
@@ -20,7 +20,8 @@ type avl = Available | Unavailable
 
 type lab_env = T.typ T.Env.t
 type ret_env = T.typ option
-type val_env  = (T.typ * Source.region * Scope.val_kind * avl) T.Env.t
+type val_info = T.typ * Source.region * Scope.val_kind * avl
+type val_env  = val_info T.Env.t
 
 (* separate maps for values and types; entries only for _public_ elements *)
 type visibility_src = {depr : string option; id_region : Source.region; field_region : Source.region}
@@ -157,11 +158,21 @@ let display_given_arg_types fmt types =
   else
     Format.fprintf fmt "But got %d argument%s of type:%a" (List.length types) (plural_typs types) (display_many display_typ_expand) types
 
-let display_obj fmt (s, fs) =
-  if !Flags.ai_errors || (List.length fs) < 16 then
-    Format.fprintf fmt "type:%a" display_typ (T.Obj(s, fs))
-  else
-    Format.fprintf fmt "%s." (String.trim(T.string_of_obj_sort s))
+let display_obj fmt typ =
+  match T.normalize typ with
+  | T.Obj(s, fs) ->
+     if !Flags.ai_errors || (List.length fs) < 16 then
+       Format.fprintf fmt "type:%a" display_typ_expand typ
+     else
+       Format.fprintf fmt "%s."
+         T.(match s with
+          | Actor -> "actor"
+          | Module -> "module"
+          | Object -> "object"
+          (* these should not occur, but let's be forgiving*)
+          | Mixin -> "mixin"
+          | Memory -> "memory")
+  | _ -> Format.fprintf fmt "type:%a" display_typ typ
 
 let display_vals fmt vals =
   if !Flags.ai_errors then
@@ -229,6 +240,8 @@ let info env at fmt =
 
 let check_deprecation env at desc id depr =
   match depr with
+  | Some ("M0235" as code) ->
+    warn env at code "%s %s is deprecated for caffeine" desc id
   | Some ("M0199" as code) ->
     if !(env.reported_stable_memory) then ()
     else begin
@@ -507,7 +520,7 @@ and check_obj_path' env path : T.typ =
     | exception Invalid_argument _ ->
       error env id.at "M0028" "field %s does not exist in %a%s"
         id.it
-        display_obj (s, fs)
+        display_obj (T.Obj(s, fs))
         (Suggest.suggest_id "field" id.it
           (List.filter_map
             (function
@@ -538,7 +551,8 @@ and check_typ_path' env path : T.con =
         c
       | exception Invalid_argument _ ->
         error env id.at "M0030" "type field %s does not exist in %a%s"
-          id.it display_obj (s, fs)
+          id.it
+          display_obj (T.Obj(s, fs))
           (Suggest.suggest_id "type field" id.it
              (List.filter_map
                (function { T.lab; T.typ=T.Typ _;_ } -> Some lab
@@ -1071,7 +1085,7 @@ let rec is_explicit_exp e =
   | RelE _ | NotE _ | AndE _ | OrE _ | ImpliesE _ | OldE _ | ShowE _ | ToCandidE _ | FromCandidE _
   | AssignE _ | IgnoreE _ | AssertE _ | DebugE _
   | WhileE _ | ForE _
-  | AnnotE _ | ImportE _ ->
+  | AnnotE _ | ImportE _ | ImplicitLibE _ ->
     true
   | LitE l -> is_explicit_lit !l
   | UnE (_, _, e1) | OptE e1 | DoOptE e1
@@ -1336,23 +1350,19 @@ type hole_candidate =
 let suggestion_of_candidate candidate =
   Option.fold ~none:candidate.desc ~some:Suggest.module_name_as_url candidate.module_name_opt
 
-(* All candidates are subtypes of the required type. The "greatest" of
-   these types is the "closest" to the required type. If we can
-   uniquely identify a single candidate that is the supertype of all
-   other candidates we pick it. *)
-let disambiguate_resolutions (candidates : hole_candidate list) =
-  let add_candidate (frontiers : hole_candidate list) (c : hole_candidate) =
-    let rec go (fs : hole_candidate list) = match fs with
+let disambiguate_resolutions (rel : 'candidate -> 'candidate -> bool) (candidates : 'candidate list) =
+  let add_candidate (frontiers : 'candidate list) (c : 'candidate) =
+    let rec go (fs : 'candidate list) = match fs with
       | [] -> [c]
       | f::fs' ->
-         if T.sub c.typ f.typ then
-           if T.eq c.typ f.typ then
+         if rel c f then
+           if rel f c then
              (* c = f, so we keep both *)
              f :: go fs'
            else
              (* c <: f, so f absorbs c *)
              fs
-         else if T.sub f.typ c.typ then
+         else if rel f c then
            (* f <: c, so c absorbs f *)
            go fs'
          else
@@ -1364,6 +1374,20 @@ let disambiguate_resolutions (candidates : hole_candidate list) =
   match List.fold_left add_candidate [] candidates with
   | [dom] -> Some dom
   | _ -> None
+
+let is_lib_module (n, t) =
+  match T.normalize t with
+  | T.Obj (T.Module, fs) -> Some (n, (t, fs))
+  | _ -> None
+
+let is_val_module (n, ((t, _, _, _) : val_info)) =
+  is_lib_module (n, t)
+
+let module_exp in_libs env module_name =
+  if not in_libs then
+    VarE {it = module_name; at = no_region; note = Const}
+  else
+    ImplicitLibE module_name
 
 (** Searches for hole resolutions for [name] on a given [hole_sort] and [typ].
     Returns [Ok(candidate)] when a single resolution is
@@ -1377,10 +1401,7 @@ let resolve_hole env at hole_sort typ =
     | Named lab1 -> lab = lab1
     | Anon _ -> true
   in
-  let is_module (n, (t, _, _, _)) = match T.normalize t with
-    | T.Obj (T.Module, fs) -> Some (n, fs)
-    | _ -> None
-  in
+
   let is_matching_typ typ1 = T.sub typ1 typ
   in
   let has_matching_field_typ = function
@@ -1391,12 +1412,12 @@ let resolve_hole env at hole_sort typ =
        then Some (lab1, typ1, src.T.region)
        else None
   in
-  let find_candidate_fields (module_name, fs) =
+  let find_candidate_fields in_libs (module_name, (_, fs)) =
     List.filter_map has_matching_field_typ fs |>
       List.map (fun (lab, typ, region)->
           let path =
             { it = DotE(
-                { it = VarE {it = module_name; at = no_region; note = Const};
+                { it = module_exp in_libs env module_name;
                   at = Source.no_region;
                   note = empty_typ_note
                 },
@@ -1425,10 +1446,10 @@ let resolve_hole env at hole_sort typ =
       List.of_seq |>
       List.partition (fun (desc : hole_candidate) -> is_matching_lab desc.id)
   in
-  let eligible_fields () =
-    T.Env.to_seq env.vals |>
-      Seq.filter_map is_module |>
-      Seq.map find_candidate_fields |>
+  let candidates in_libs xs f =
+    T.Env.to_seq xs |>
+      Seq.filter_map f |>
+      Seq.map (find_candidate_fields in_libs) |>
       List.of_seq |>
       List.flatten |>
       List.partition (fun (desc : hole_candidate) -> is_matching_lab desc.id)
@@ -1437,7 +1458,7 @@ let resolve_hole env at hole_sort typ =
     match eligible_ids with
     | [id] -> ([id], []) (* first look in local env, otherwise consider module entries *)
     | _ ->
-       let (eligible_fields, explicit_fields) = eligible_fields () in
+       let (eligible_fields, explicit_fields) = candidates false env.vals is_val_module in
        (eligible_ids @ eligible_fields,
         explicit_ids @ explicit_fields)
   in
@@ -1466,25 +1487,22 @@ let resolve_hole env at hole_sort typ =
              candidate.desc mid id mod_desc mid id call_region call_src)
       explicit_terms
   in
+  (* All candidates are subtypes of the required type. The "greatest" of these types is the "closest" to the required type.
+     If we can uniquely identify a single candidate that is the supertype of all other candidates we pick it. *)
+  let disambiguate_holes = disambiguate_resolutions (fun (c1 : hole_candidate) c2 -> T.sub c1.typ c2.typ) in
   match eligible_terms with
   | [term] -> Ok term
   | [] ->
-     let (lib_terms, _) =
-       T.Env.to_seq env.libs |>
-         Seq.filter_map (fun (n, t) ->
-             match t with
-             | T.Obj (T.Module, fs) -> Some (n, fs)
-             | _ -> None) |>
-         Seq.map find_candidate_fields |>
-         List.of_seq |>
-         List.flatten |>
-         List.partition (fun (candidate : hole_candidate) -> is_matching_lab candidate.id)
-     in
-     Error (List.map suggestion_of_candidate lib_terms,
-            List.map (fun candidate -> candidate.desc) explicit_terms,
-            renaming_hints)
+    let (lib_terms, _) = candidates true env.libs is_lib_module in
+    (match if Option.is_some !Flags.implicit_package then disambiguate_holes lib_terms else None with
+    | Some term -> Ok term
+    | None ->
+      Error (List.map suggestion_of_candidate lib_terms,
+              List.map (fun candidate -> candidate.desc) explicit_terms,
+              renaming_hints)
+    )
   | terms -> begin
-     match disambiguate_resolutions terms with
+     match disambiguate_holes terms with
      | Some term -> Ok term
      | None ->
        let terms = List.map (fun term -> term.desc) terms in
@@ -1500,8 +1518,10 @@ let resolve_hole env at hole_sort typ =
 
 type ctx_dot_candidate =
   { module_name : T.lab;
+    path : exp;
+    arg_ty : T.typ;
     func_ty : T.typ;
-    module_ty : T.typ;
+    inst : T.typ list;
   }
 
 (** Searches for contextual resolutions for [name] on a given
@@ -1510,54 +1530,108 @@ type ctx_dot_candidate =
     matching module could be imported, and reports an ambiguity error
     when finding multiple resolutions.
  *)
+
+(* Does an instantiation for [tbs] exist that makes [t1] <: [t2]? *)
+let permissive_sub t1 (tbs, t2) =
+  try
+    let (inst, c) = Bi_match.bi_match_subs None tbs None [t1, t2] [] in
+    ignore (Bi_match.finalize inst c []);
+    Some inst
+  with _ ->
+    None
+
+let has_matching_self receiver_ty tf = match tf with
+  | T.{ lab = "Self"; typ = T.Typ con; _ } ->
+     (match Cons.kind con with
+      | T.Def(tbs, t') -> permissive_sub receiver_ty (tbs, t') <> None
+      | _ -> false)
+  | _ -> false
+
 let contextual_dot env name receiver_ty =
-  (* Does an instantiation for [tbs] exist that makes [t1] <: [t2]? *)
-  let permissive_sub t1 (tbs, t2) =
-    try
-      let (s, c) = Bi_match.bi_match_subs None tbs None [t1, t2] [] in
-      ignore (Bi_match.finalize s c []);
-      true
-    with _ ->
-      false in
-  let is_module (n, (t, _, _, _)) = match T.normalize t with
-    | T.Obj (T.Module, fs) -> Some (n, (t, fs))
-    | _ -> None in
-  let has_matching_self tf = match tf with
-    | T.{ lab = "Self"; typ = T.Typ con; _ } ->
-       (match Cons.kind con with
-       | T.Def(tbs, t') -> permissive_sub receiver_ty (tbs, t')
-       | _ -> false)
-    | _ -> false in
+  let has_matching_self tf = has_matching_self receiver_ty tf in
   let has_matching_self_type (_, (_, fs)) = List.exists has_matching_self fs in
-  let is_matching_func = function
-    | T.{ lab; typ = T.Func (_, _, tbs, first_arg::_, _) as typ; _ } when lab = name.it ->
-      if permissive_sub receiver_ty (tbs, first_arg) then Some typ else None
+  let is_matching_func field =
+    if not (String.equal field.T.lab name.it) then None
+    else match T.normalize field.T.typ with
+    | T.Func (_, _, tbs, first_arg::_, _) as typ ->
+      (match permissive_sub receiver_ty (tbs, first_arg) with
+        | Some inst -> Some (T.open_ inst first_arg, typ, inst)
+        | _ -> None)
     | _ -> None in
-  let find_candidate (module_name, (module_ty, fs)) =
+  let find_candidate in_libs (module_name, (module_ty, fs)) =
     List.find_map is_matching_func fs |>
-      Option.map (fun func_ty -> { module_name; func_ty; module_ty }) in
-  let eligible_funcs =
-    T.Env.to_seq env.vals |>
-      Seq.filter_map is_module |>
+      Option.map (fun (arg_ty, func_ty, inst) ->
+        let path = {
+          it = DotE({
+              it = module_exp in_libs env module_name;
+              at = name.at;
+              note = empty_typ_note
+            }, name, ref None);
+          at = name.at;
+          note = empty_typ_note }
+        in
+        { module_name; path; func_ty; arg_ty; inst }) in
+  let candidates in_libs xs f =
+    T.Env.to_seq xs |>
+      Seq.filter_map f |>
       Seq.filter has_matching_self_type |>
-      Seq.filter_map find_candidate |>
+      Seq.filter_map (find_candidate in_libs) |>
       List.of_seq in
-  match eligible_funcs with
-  | [oc] -> Ok oc
+  (* All candidate functions accept supertypes of the required type as their first arguments.
+     The "smallest" of these types is the closest to the required type. *)
+  let disambiguate_candidates = disambiguate_resolutions (fun c1 c2 -> T.sub c2.arg_ty c1.arg_ty) in
+  match candidates false env.vals is_val_module with
+  | [c] -> Ok c
   | [] ->
-     let lib_candidates =
-       T.Env.to_seq env.libs |>
-         Seq.filter_map (fun (n, t) ->
-             match t with
-             | T.Obj (T.Module, fs) -> Some (n, (t, fs))
-             | _ -> None) |>
-         Seq.filter has_matching_self_type |>
-         Seq.filter_map find_candidate |>
-         List.of_seq in
-     Error (List.map (fun candidate -> Suggest.module_name_as_url candidate.module_name) lib_candidates)
-  | ocs ->
-     let candidates = List.map (fun oc -> oc.module_name) ocs in
-     error env name.at "M0224" "overlapping resolution for `%s` in scope from these modules: %s" name.it (String.concat ", " candidates)
+    (match candidates true env.libs is_lib_module with
+    | [c] when Option.is_some !Flags.implicit_package -> Ok c
+    | lib_candidates ->
+      match if Option.is_some !Flags.implicit_package then disambiguate_candidates lib_candidates else None with
+      | Some c -> Ok c
+      | None ->  Error (List.map (fun candidate -> Suggest.module_name_as_url candidate.module_name) lib_candidates))
+  | cs -> match disambiguate_candidates cs with
+    | Some c -> Ok c
+    | None ->
+      let candidates = List.map (fun c -> c.module_name) cs in
+      error env name.at "M0224" "overlapping resolution for `%s` in scope from these modules: %s" name.it (String.concat ", " candidates)
+
+let check_can_dot env ctx_dot (exp : Syntax.exp) tys es at =
+  if not env.pre then
+  if Flags.get_warning_level "M0236" <> Flags.Allow then
+  match ctx_dot with
+  | Some _ -> () (* already dotted *)
+  | None ->
+    match exp.it, tys, es with
+    | (DotE(obj_exp, id, _), receiver_ty :: tys, e::es) ->
+      begin
+        if (id.it = "equal" || Lib.String.chop_prefix "compare" id.it <> None) && List.length tys = 1 then ()
+        else
+          let quote e =
+            if e.at.left.line <> e.at.right.line then "..." else
+            match Source.read_region e.at with
+            | None -> "..."
+            | Some s ->
+               match e.it with
+               | VarE _ | CallE _ | DotE _ -> s
+               | e -> "("^s^")"
+          in
+          match contextual_dot env id receiver_ty with
+          | Error _ -> ()
+          | Ok {path;_} ->
+            (match path.it, exp.it with
+             | DotE ({ it = VarE {it = mod_id0; _};_ },
+                     { it = id0; _},
+                    _),
+               DotE ({ it = VarE {it = mod_id1; note = Const; _};_ },
+                     { it = id1; _},
+                     _)  when mod_id0 = mod_id1 && id0 = id1 ->
+                warn env at "M0236" "You can use the dot notation `%s.%s(...)` here"
+                  (quote e)
+                  id.it
+             | _ -> ())
+      end
+    | _, _, _ -> ()
+
 
 let rec infer_exp env exp : T.typ =
   infer_exp' T.as_immut env exp
@@ -1566,13 +1640,17 @@ and infer_exp_mut env exp : T.typ =
   infer_exp' Fun.id env exp
 
 and infer_exp_promote env exp : T.typ =
+  snd (infer_exp_and_promote env exp)
+
+and infer_exp_and_promote env exp : T.typ * T.typ =
   let t = infer_exp env exp in
   let t' = T.promote t in
   if t' = T.Pre then
     error env exp.at "M0053"
       "cannot infer type of expression while trying to infer surrounding class type,\nbecause its type is a forward reference to type%a"
       display_typ_expand t;
-  t'
+  t, t'
+
 
 and infer_exp_wrapper inf f env exp : T.typ =
   assert (exp.note.note_typ = T.Pre);
@@ -1759,9 +1837,10 @@ and infer_exp'' env exp : T.typ =
   | ObjE (exp_bases, exp_fields) ->
     infer_check_bases_fields env [] exp.at exp_bases exp_fields
   | DotE (exp1, id, _) ->
-    (match try_infer_dot_exp env exp.at exp1 id with
+    (match try_infer_dot_exp env exp.at exp1 id ("", (fun dot_typ -> true))  with
     | Ok t -> t
-    | Error (_, e) ->
+    | Error (_, mk_e) ->
+      let e = mk_e() in
       Diag.add_msg env.msgs e;
       raise Recover)
   | AssignE (exp1, exp2) ->
@@ -2064,6 +2143,10 @@ and infer_exp'' env exp : T.typ =
     T.unit
   | ImportE (f, ri) ->
     check_import env exp.at f ri
+  | ImplicitLibE lib ->
+    match T.Env.find_opt lib env.libs with
+    | Some t -> t
+    | None -> failwith "ImplicitLibE not found in env.libs"
 
 and infer_bin_exp env exp1 exp2 =
   match is_explicit_exp exp1, is_explicit_exp exp2 with
@@ -2083,38 +2166,57 @@ and infer_bin_exp env exp1 exp2 =
 (* Returns `Ok` when finding an object with a matching field or
    `Error` with the type of the receiver as well as the error message
    to report. This is used to delay the reporting for contextual dot resulution *)
-and try_infer_dot_exp env at exp id =
-  let t1 = infer_exp_promote env exp in
+and try_infer_dot_exp env at exp id (desc, pred) =
+  let t0, t1 = infer_exp_and_promote env exp in
   let fields =
     try Ok(T.as_obj_sub [id.it] t1) with Invalid_argument _ ->
     try Ok(array_obj (T.as_array_sub t1)) with Invalid_argument _ ->
     try Ok(blob_obj (T.as_prim_sub T.Blob t1)) with Invalid_argument _ ->
     try Ok(text_obj (T.as_prim_sub T.Text t1)) with Invalid_argument _ ->
-      Error(t1, type_error exp.at "M0070" (Format.asprintf
-              "expected object type, but expression produces type%a"
-              display_typ_expand t1))
+      Error(t1, fun () ->
+        type_error exp.at "M0070"
+          (Format.asprintf
+             "expected object type, but expression produces type%a"
+             display_typ_expand t0))
   in
   match fields with
   | Error e -> Error e
   | Ok((s, tfs)) -> begin
+    let suggest () =
+      Suggest.suggest_id "field" id.it
+        (List.filter_map
+           (function
+              { T.typ=T.Typ _;_} -> None
+            | {T.lab;_} -> Some lab) tfs)
+    in
     match T.lookup_val_field id.it tfs with
     | T.Pre ->
       error env at "M0071"
         "cannot infer type of forward field reference %s"
         id.it
-    | t ->
+    | t when pred t ->
       if not env.pre then
         check_deprecation env at "field" id.it (T.lookup_val_deprecation id.it tfs);
       Ok(t)
+    | t (* when not (pred t) *) ->
+      Error(t1, fun () ->
+        type_error id.at "M0234"
+          (Format.asprintf "field %s does exist in %a\nbut is not %s.\n%s"
+             id.it
+             display_obj t0
+             desc
+             (suggest ())))
     | exception Invalid_argument _ ->
-      Error(t1, type_error id.at "M0072" (Format.asprintf "field %s does not exist in %a%s"
-          id.it
-          display_obj (s, tfs)
-          (Suggest.suggest_id "field" id.it
-             (List.filter_map
-                (function
-                   { T.typ=T.Typ _;_} -> None
-                 | {T.lab;_} -> Some lab) tfs))))
+      Error(t1, fun () ->
+        type_error id.at "M0072"
+          (Format.asprintf "field %s does not exist in %a%s"
+             id.it
+             display_obj t0
+             (Suggest.suggest_id "field" id.it
+                (List.filter_map
+                   (function
+                      { T.typ=T.Typ _;_} -> None
+                    | {T.lab;_} -> Some lab) tfs))))
     end
 
 and infer_exp_field env rf =
@@ -2235,25 +2337,28 @@ and check_exp' env0 t exp : T.typ =
       check_exp env t path;
       t
     | Error (import_suggestions, explicit_suggestions, renaming_hints) ->
-      let import_sug =
-        if import_suggestions = [] then
-           let desc = match s with Named id -> " named " ^ quote id | _ -> "" in
-           Format.sprintf
+      if not env.pre then begin
+        let import_sug =
+          if import_suggestions = [] then
+            let desc = match s with Named id -> " named " ^ quote id | _ -> "" in
+            Format.sprintf
              "\nHint: If you're trying to omit an implicit argument%s you need to have a matching declaration%s in scope."
              desc desc
-         else Format.sprintf "\nHint: Did you mean to import %s?" (String.concat " or " import_suggestions)
-      in
-      let explicit_sug =
-         if explicit_suggestions = [] then ""
-         else Format.sprintf "\nHint: Did you mean to explicitly use %s?" (String.concat " or " explicit_suggestions)
-      in
-      renaming_hints env;
-      error env exp.at "M0230" "Cannot determine implicit argument %s of type%a%s%s"
-        (desc s)
-        display_typ t
-        import_sug
-        explicit_sug
-    end
+          else Format.sprintf "\nHint: Did you mean to import %s?" (String.concat " or " import_suggestions)
+        in
+        let explicit_sug =
+          if explicit_suggestions = [] then ""
+          else Format.sprintf "\nHint: Did you mean to explicitly use %s?" (String.concat " or " explicit_suggestions)
+        in
+        renaming_hints env;
+        local_error env exp.at "M0230" "Cannot determine implicit argument %s of type%a%s%s"
+          (desc s)
+          display_typ t
+          import_sug
+          explicit_sug
+      end;
+      t
+  end
   | PrimE s, T.Func _ ->
     if not !Flags.wasm_components && Lib.String.chop_prefix "component:" s <> None then
       error env exp.at "M0234" "Wasm components support not enabled, use -wasm-components flag";
@@ -2415,7 +2520,7 @@ and check_exp' env0 t exp : T.typ =
     let {T.typ; _} = List.find (fun T.{lab; typ;_} -> lab = id.it) fs in
     check_exp env typ exp1;
     t
-  | ImportE _, t ->
+  | (ImportE _ | ImplicitLibE _), t ->
     t
   | e, _ ->
     let t' = infer_exp env0 exp in
@@ -2522,31 +2627,29 @@ and detect_lost_fields env t = function
   | _ -> ()
 
 and infer_callee env exp =
+  let is_func_typ typ = T.(
+    match promote typ with
+    | Func _ | Non -> true
+    | _ -> false)
+  in
   match exp.it with
   | DotE(exp1, id, note) -> begin
-    match try_infer_dot_exp env exp.at exp1 id with
-    | Ok t -> infer_exp_wrapper (fun _ _ -> t) T.as_immut env exp, None
-    | Error (t1, e) ->
+    match try_infer_dot_exp env exp.at exp1 id ("a function", is_func_typ) with
+    | Ok t ->
+      infer_exp_wrapper (fun _ _ -> t) T.as_immut env exp, None
+    | Error (t1, mk_e) ->
       match contextual_dot env id t1 with
       | Error [] ->
-         let sug = "\nHint: If you're trying to use a contextual call you need to import the corresponding module." in
-         Diag.add_msg env.msgs Diag.{ e with text = e.text ^ sug }; raise Recover
+         Diag.add_msg env.msgs (mk_e ()); raise Recover
       | Error suggestions ->
+         let e = mk_e () in
          let sug = Format.sprintf "\nHint: Did you mean to import %s?" (String.concat " or " suggestions) in
          Diag.add_msg env.msgs Diag.{ e with text = e.text ^ sug }; raise Recover
-      | Ok { module_name; module_ty; func_ty } ->
-         if not env.pre then
-         use_identifier env module_name;
-         note := Some {
-           it = DotE({
-               it = VarE { it = module_name; at = no_region; note = Const };
-               at = id.at;
-               note = { note_eff = T.Triv; note_typ = module_ty }
-             }, id, ref None);
-           at = id.at;
-           note = { note_eff = T.Triv; note_typ = func_ty }
-         };
-         func_ty, Some (exp1, t1)
+      | Ok { module_name; path; func_ty; inst; _ } ->
+        note := Some path;
+        if not env.pre then
+          check_exp env func_ty path;
+        func_ty, Some (exp1, t1, id.it, inst)
      end
   | _ ->
      infer_exp_promote env exp, None
@@ -2596,6 +2699,36 @@ and insert_holes at ts es =
   | [arg] -> arg.it
   | args -> TupE args
 
+and check_explicit_arguments env saturated_arity implicits_arity arg_typs syntax_args =
+    if Flags.get_warning_level "M0237" <> Flags.Allow then
+      if List.length syntax_args = saturated_arity && implicits_arity < saturated_arity then
+        let _, explicit_implicits = List.fold_right2
+          (fun typ arg (pos, acc) ->
+             pos + 1,
+             match as_implicit typ with
+             | None -> acc
+             | Some name ->
+                match resolve_hole env arg.at (match name with "_" -> Anon pos | id -> Named id) typ with
+                | Error _ -> acc
+                | Ok {path;_} ->
+                   match path.it, arg.it with
+                   | VarE {it = id0; _},
+                     VarE {it = id1; note = Const; _}
+                        when id0 = id1 ->
+                      (id1, arg) :: acc
+                   | DotE ({ it = VarE {it = mod_id0; _};_ },
+                           { it = id0; _},
+                           _),
+                     DotE ({ it = VarE {it = mod_id1; note = Const; _};_ },
+                           { it = id1; _},
+                           _) when mod_id0 = mod_id1 && id0 = id1 ->
+                      (mod_id1 ^ "." ^ id1, arg) :: acc
+                   | _ -> acc)
+          arg_typs syntax_args (0, [])
+        in
+        if (List.length explicit_implicits) = saturated_arity - implicits_arity then
+          List.iter (fun (name, exp) -> warn env exp.at "M0237" "The `%s` argument can be inferred and omitted here (the function parameter is `implicit`)." name)  explicit_implicits
+
 and infer_call env exp1 inst (parenthesized, ref_exp2) at t_expect_opt =
   let exp2 = !ref_exp2 in
   let n = match inst.it with None -> 0 | Some (_, typs) -> List.length typs in
@@ -2620,8 +2753,9 @@ and infer_call env exp1 inst (parenthesized, ref_exp2) at t_expect_opt =
     | _ -> [exp2]
   in
   let t_args, extra_subtype_problems = match ctx_dot with
-    | None -> t_args, []
-    | Some(e, t) -> begin
+    | None ->
+      t_args, []
+    | Some(e, t, _id, _inst) -> begin
       match t_args with
       | t'::ts -> ts, [(t, t')]
       | [] -> assert false
@@ -2660,11 +2794,11 @@ and infer_call env exp1 inst (parenthesized, ref_exp2) at t_expect_opt =
       if not env.pre then check_exp_strong env t_arg' exp2
       else if typs <> [] && Flags.is_warning_enabled "M0223" &&
         is_redundant_instantiation ts env (fun env' ->
-          infer_call_instantiation env' t1 tbs t_arg t_ret exp2 at t_expect_opt extra_subtype_problems) then
+          infer_call_instantiation env' t1 ctx_dot tbs t_arg t_ret exp2 at t_expect_opt extra_subtype_problems) then
             warn env inst.at "M0223" "redundant type instantiation";
       ts, t_arg', t_ret'
     | _::_, None -> (* implicit, infer *)
-      infer_call_instantiation env t1 tbs t_arg t_ret exp2 at t_expect_opt extra_subtype_problems
+      infer_call_instantiation env t1 ctx_dot tbs t_arg t_ret exp2 at t_expect_opt extra_subtype_problems
   in
   inst.note <- ts;
   if not env.pre then begin
@@ -2678,12 +2812,15 @@ and infer_call env exp1 inst (parenthesized, ref_exp2) at t_expect_opt =
           "shared function call result contains abstract type%a"
           display_typ_expand t_ret';
     end;
-    match T.(is_shared_sort sort || is_async t_ret'), inst.it, tbs with
+    begin match T.(is_shared_sort sort || is_async t_ret'), inst.it, tbs with
     | false, Some (true, _), ([] | T.{ sort = Type; _ } :: _) ->
        local_error env inst.at "M0196" "unexpected `system` capability (try deleting it)"
     | false, (None | Some (false, _)), T.{ sort = Scope; _ } :: _ ->
        warn env at "M0195" "this function call implicitly requires `system` capability and may perform undesired actions (please review the call and provide a type instantiation `<system%s>` to suppress this warning)" (if List.length tbs = 1 then "" else ", ...")
     | _ -> ()
+    end;
+    check_can_dot env ctx_dot exp1 (List.map (T.open_ ts) t_args) syntax_args at;
+    check_explicit_arguments env saturated_arity implicits_arity (List.map (T.open_ ts) t_args) syntax_args;
   end;
   (* note t_ret' <: t checked by caller if necessary *)
   t_ret'
@@ -2704,7 +2841,7 @@ and wrong_call_args env tbs at t_args implicits_arity syntax_args =
     display_expected_arg_types expected_types
     display_given_arg_types given_types
 
-and infer_call_instantiation env t1 tbs t_arg t_ret exp2 at t_expect_opt extra_subtype_problems =
+and infer_call_instantiation env t1 ctx_dot tbs t_arg t_ret exp2 at t_expect_opt extra_subtype_problems =
   (*
   Partial Argument Inference:
   We need to infer the type of the argument and find the best instantiation for the call expression.
@@ -2754,7 +2891,11 @@ and infer_call_instantiation env t1 tbs t_arg t_ret exp2 at t_expect_opt extra_s
   if Bi_match.debug then debug_print_infer_defer_split exp2 t_arg t2 subs deferred;
 
   (* In case of an early error, we need to replace Type.Var with Type.Con for a better error message *)
-  let err_ts = ref None in
+  let err_ts = ref
+      (match ctx_dot with
+      | Some (_, _, _, ts) -> Some ts
+      | _ -> None)
+  in
   let err_subst t =
     let ts = match !err_ts with
       | None -> T.open_binds tbs
@@ -2832,22 +2973,72 @@ and infer_call_instantiation env t1 tbs t_arg t_ret exp2 at t_expect_opt extra_s
       fix (T.open_ ts) to_fix;
       fix (T.open_ ts) deferred;
     end;
-(*
+  (*
     if not env.pre then
       info env at "inferred instantiation <%s>"
         (String.concat ", " (List.map T.string_of_typ ts));
   *)
     ts, T.open_ ts t_arg, T.open_ ts t_ret
   with Bi_match.Bimatch msg ->
-    error env at "M0098"
-      "cannot implicitly instantiate function of type%a\nto argument of type%a%s\nbecause %s"
-      display_typ t1
-      display_typ (err_subst t2)
-      (match t_expect_opt with
-        | None -> ""
-        | Some t ->
-          Format.asprintf "\nto produce result of type%a" display_typ t)
-      msg
+    let t1 = match T.normalize t1 with
+      | T.Func(s, c, tbs, ts1, ts2) ->
+        T.Func(s, c, [], List.map err_subst ts1, List.map err_subst ts2)
+      | t1 -> t1
+    in
+    let remove_holes_nary ts =
+      match exp2.it, ts with
+        HoleE _, [_] ->
+        ts
+      | TupE es, ts when List.length es = List.length ts ->
+        let ets = List.combine es ts in
+        List.filter_map (fun (e, t) ->
+          match e.it with
+          | HoleE _ -> None
+          | _ -> Some t)
+          ets
+      | e -> ts
+    in
+    let strip_receiver ty = match ty with
+      | T.Func(s, c, tbs, t1::ts1, ts2) ->
+        T.Func(s, c, tbs, ts1, ts2)
+      |  _ -> ty
+    in
+    let desc, t1'  = match ctx_dot with
+      | None -> "function", t1
+      | Some (_, receiver_ty, id, _) ->
+        Printf.sprintf "function `.%s`" id,
+        strip_receiver t1
+    in
+    let t1'' = match T.normalize t1' with
+      | T.Func(s, c, tbs, ts1, ts2) ->
+        T.Func(s, c, [], remove_holes_nary ts1, List.map err_subst ts2)
+      | t1 -> t1
+    in
+    let remove_holes typ =
+      T.seq (remove_holes_nary (match typ with T.Tup ts -> ts | t -> [t]))
+    in
+    let t2' = remove_holes t2 in
+    if Bi_match.debug then
+      error env at "M0098"
+        "cannot implicitly instantiate %s of type%a\nto argument of type%a%s\nbecause %s"
+        desc
+        display_typ t1
+        display_typ (err_subst t2')
+        (match t_expect_opt with
+         | None -> ""
+         | Some t ->
+           Format.asprintf "\nto produce result of expected type%a" display_typ t)
+        msg
+    else
+      error env at "M0098"
+        "cannot apply %s of type%a\nto argument of type%a%s"
+        desc
+        display_typ t1''
+        display_typ (err_subst t2')
+        (match t_expect_opt with
+         | None -> ""
+         | Some t ->
+           Format.asprintf "\nto produce result of expected type%a" display_typ t)
 
 and is_redundant_instantiation ts env infer_instantiation =
   assert env.pre;
@@ -3945,6 +4136,10 @@ and infer_val_path env exp : T.typ option =
     (match T.Env.find_opt id.it env.vals with (* TBR: return None for Unavailable? *)
      | Some (t, _, _, _) -> Some t
      | _ -> None)
+  | ImplicitLibE lib ->
+    (match T.Env.find_opt lib env.libs with
+    | Some t -> Some t
+    | None -> None)
   | DotE (path, id, _) ->
     (match infer_val_path env path with
      | None -> None